Re-organize the analyzer classes ready to introduce graphs
