+++ src/analyze/util/visitor.py
